+++ asyncprawcore/sessions.py
@@ -11,15 +11,15 @@ from contextlib import asynccontextmanager
 from copy import deepcopy
 from dataclasses import dataclass
 from pprint import pformat
-from typing import IO, TYPE_CHECKING, Any
+from typing import IO, TYPE_CHECKING, Any, ClassVar
 from urllib.parse import urljoin
 
 from aiohttp.web import HTTPRequestTimeout
 
-from .auth import BaseAuthorizer
-from .codes import codes
-from .const import TIMEOUT, WINDOW_SIZE
-from .exceptions import (
+from asyncprawcore.auth import BaseAuthorizer
+from asyncprawcore.codes import codes
+from asyncprawcore.const import TIMEOUT, WINDOW_SIZE
+from asyncprawcore.exceptions import (
     BadJSON,
     BadRequest,
     Conflict,
@@ -35,8 +35,8 @@ from .exceptions import (
     UnavailableForLegalReasons,
     URITooLong,
 )
-from .rate_limit import RateLimiter
-from .util import authorization_error_class
+from asyncprawcore.rate_limit import RateLimiter
+from asyncprawcore.util import authorization_error_class
 
 if TYPE_CHECKING:
     from collections.abc import AsyncGenerator, Mapping
@@ -44,7 +44,7 @@ if TYPE_CHECKING:
     from aiohttp import ClientResponse
     from typing_extensions import Self
 
-    from .requestor import Requestor
+    from asyncprawcore.requestor import Requestor
 
 log = logging.getLogger(__package__)
 
@@ -83,7 +83,7 @@ class RetryStrategy(ABC):
 class FiniteRetryStrategy(RetryStrategy):
     """A ``RetryStrategy`` that retries requests a finite number of times."""
 
-    DEFAULT_RETRIES = 2
+    DEFAULT_RETRIES: ClassVar[int] = 2
 
     retries: int = DEFAULT_RETRIES
 
@@ -106,7 +106,7 @@ class Session:
     """The low-level connection interface to Reddit's API."""
 
     RETRY_EXCEPTIONS = (ConnectionError, HTTPRequestTimeout)
-    RETRY_STATUSES = {
+    RETRY_STATUSES: ClassVar = {
         520,
         522,
         codes["bad_gateway"],
@@ -115,7 +115,7 @@ class Session:
         codes["request_timeout"],
         codes["service_unavailable"],
     }
-    STATUS_EXCEPTIONS = {
+    STATUS_EXCEPTIONS: ClassVar = {
         codes["bad_gateway"]: ServerError,
         codes["bad_request"]: BadRequest,
         codes["conflict"]: Conflict,
@@ -137,7 +137,7 @@ class Session:
         520: ServerError,
         522: ServerError,
     }
-    SUCCESS_STATUSES = {codes["accepted"], codes["created"], codes["ok"]}
+    SUCCESS_STATUSES: ClassVar = {codes["accepted"], codes["created"], codes["ok"]}
 
     @staticmethod
     def _log_request(
@@ -223,12 +223,12 @@ class Session:
             retry_strategy_state=retry_strategy_state.consume_available_retry(),
             timeout=timeout,
             url=url,
-            # noqa: E501
         )
 
     @asynccontextmanager
     async def _make_request(
         self,
+        *,
         data: list[tuple[str, object]] | bytes | IO[Any] | str | None,
         json: dict[str, object] | list[object] | None,
         method: str,
@@ -372,7 +372,7 @@ class Session:
                 except ValueError:
                     raise BadJSON(response) from None
         except RequestException as exception:
-            if retry_strategy_state.should_retry_on_failure() and isinstance(  # noqa: E501
+            if retry_strategy_state.should_retry_on_failure() and isinstance(
                 exception.original_exception, self.RETRY_EXCEPTIONS
             ):
                 return await self._do_retry(
@@ -399,24 +399,25 @@ class Session:
 
     async def request(
         self,
-        method: str,
-        path: str,
+        *,
         data: dict[str, object] | bytes | IO[Any] | str | None = None,
         files: dict[str, IO[Any]] | None = None,
         json: dict[str, object] | list[object] | None = None,
+        method: str,
         params: Mapping[str, object] | None = None,
+        path: str,
         timeout: float = TIMEOUT,
     ) -> dict[str, object] | str | None:
         """Return the json content from the resource at ``path``.
 
-        :param method: The request verb. E.g., ``"GET"``, ``"POST"``, ``"PUT"``.
-        :param path: The path of the request. This path will be combined with the
-            ``oauth_url`` of the Requestor.
         :param data: Dictionary, bytes, or file-like object to send in the body of the
             request.
         :param files: Dictionary, mapping ``filename`` to file-like object.
         :param json: Object to be serialized to JSON in the body of the request.
+        :param method: The request verb. E.g., ``"GET"``, ``"POST"``, ``"PUT"``.
         :param params: The query parameters to send with the request.
+        :param path: The path of the request. This path will be combined with the
+            ``oauth_url`` of the Requestor.
         :param timeout: Specifies a particular timeout, in seconds.
 
         Automatically refreshes the access token if it becomes invalid and a refresh