+++ asyncprawcore/util.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-from .exceptions import Forbidden, InsufficientScope, InvalidToken
+from asyncprawcore.exceptions import Forbidden, InsufficientScope, InvalidToken
 
 if TYPE_CHECKING:
     from aiohttp import ClientResponse