+++ tests/utils.py
@@ -62,8 +62,7 @@ class CustomPersister(FilesystemPersister):
     def load_cassette(cls, cassette_path, serializer):  # pragma: no cover
         """Load cassette."""
         try:
-            with Path(cassette_path).open() as f:
-                cassette_content = f.read()
+            cassette_content = Path(cassette_path).read_text()
         except OSError as error:
             msg = "Cassette not found."
             raise ValueError(msg) from error
@@ -85,8 +84,7 @@ class CustomPersister(FilesystemPersister):
         dirname = cassette_path.parent
         if dirname and not dirname.exists():
             dirname.mkdir(parents=True)
-        with cassette_path.open("w") as f:
-            f.write(data)
+        cassette_path.write_text(data)
 
 
 class CustomSerializer:
@@ -94,8 +92,7 @@ class CustomSerializer:
 
     @staticmethod
     def _serialize_file(file):  # pragma: no cover
-        with Path(file.name).open("rb") as f:
-            return f.read().decode("utf-8", "replace")
+        return Path(file.name).read_bytes().decode("utf-8", "replace")
 
     @staticmethod
     def deserialize(cassette_string):